Update buildAndGet script

+++ .builtJars/buildAndGet.bat
@@ -1,7 +1,7 @@
 @echo off
 
 cd /D %~dp0
-del /q *.jar >nul
+if exist *.jar del /q *.jar >nul
 
 echo Building...
 cd /D ../
@@ -16,9 +16,10 @@ move ..\forge\build\libs\*.jar . >nul
 move ..\neoforge\build\libs\*.jar . >nul
 
 echo Deleting files...
-del /q *slim.jar >nul
-del /q *sources.jar >nul
-del /q *fat.jar >nul
+if exist *dev-shadow.jar del /q *dev-shadow.jar >nul
+if exist *slim.jar del /q *slim.jar >nul
+if exist *sources.jar del /q *sources.jar >nul
+if exist *fat.jar del /q *fat.jar >nul
 
 echo Done.
 pause >nul
